seperate yt checking as an action file
